+++ CarShop/DataAccessInterface/ISalesAccessor.cs
@@ -9,7 +9,7 @@ namespace DataAccessInterface
 {
     public interface ISalesAccessor 
     {
-        int CreateSale(int EmployeeID, int CarID, int CustomerID, DateTime SaleDate, float SalePrice);
+        int CreateSale(int EmployeeID, int CarID, int CustomerID, DateTime SaleDate, Double SalePrice);
         List<SalesVM> ViewSales();
         List<SalesVM> ViewSalesForEmployee(int EmployeeID);
     }
